move to new docs page component
