+++ website/package.json
@@ -8,7 +8,7 @@
     "@hashicorp/react-alert-banner": "^4.1.0",
     "@hashicorp/react-button": "^3.0.2",
     "@hashicorp/react-content": "5.2.1",
-    "@hashicorp/react-docs-page": "6.3.1",
+    "@hashicorp/react-docs-page": "^6.3.1-canary.1",
     "@hashicorp/react-global-styles": "4.6.1",
     "@hashicorp/react-hashi-stack-menu": "^1.0.8",
     "@hashicorp/react-head": "1.1.3",
