+++ website/pages/commands/[[...page]].jsx
@@ -1,21 +1,17 @@
-import {
-  productName,
-  productSlug,
-} from 'data/metadata'
+import { productName, productSlug } from 'data/metadata'
 import order from 'data/commands-navigation.js'
-import DocsPage from 'components/temporary_docs-page'
+import DocsPage from '@hashicorp/react-docs-page'
 import {
   generateStaticPaths,
   generateStaticProps,
-} from 'components/temporary_docs-page/server'
+} from '@hashicorp/react-docs-page/server'
 
 const subpath = 'commands'
 
-function DocsLayout(props) {
+function CommandsLayout(props) {
   return (
     <DocsPage
-      productName={productName}
-      productSlug={productSlug}
+      product={{ name: productName, slug: productSlug }}
       subpath={subpath}
       order={order}
       staticProps={props}
@@ -31,4 +27,4 @@ export async function getStaticProps({ params }) {
   return generateStaticProps(subpath, productName, params)
 }
 
-export default DocsLayout
+export default CommandsLayout
