+++ website/pages/docs/[[...page]].jsx
@@ -1,24 +1,22 @@
-import { 
-  productName,
-  productSlug,
-} from 'data/metadata'
+import { productName, productSlug } from 'data/metadata'
 import order from 'data/docs-navigation.js'
-import DocsPage from 'components/temporary_docs-page'
+import DocsPage from '@hashicorp/react-docs-page'
 import {
   generateStaticPaths,
   generateStaticProps,
-} from 'components/temporary_docs-page/server'
+} from '@hashicorp/react-docs-page/server'
+import Placement from 'components/placement-table'
 
 const subpath = 'docs'
 
 function DocsLayout(props) {
   return (
     <DocsPage
-      productName={productName}
-      productSlug={productSlug}
+      product={{ name: productName, slug: productSlug }}
       subpath={subpath}
       order={order}
       staticProps={props}
+      additionalComponents={{ Placement }}
     />
   )
 }
